Fix SCTP channel close.
Bug scenario:
- Create data channels DC1 between PeerA and PeerB
- Create data channels DC2 between PeerA and PeerB
- Transfer some data using DC2
- Close DC2 from Peer A (both: In and Out channels)
- Send data on DC1 from PeerB to PeerA.
- Sending data on DC1 fails, because there is a pending reset operation

What actually happened, is that the ReConfigurationResponseParameter was not matched correctly with the previously sent OutgoingSSNResetRequestParameter (was matched with a ReConfigurationResponseParameter having the same SequenceNumber). This caused one of the peers having a pending reset operation, and sending data on other data channels were failing.
This happened only when PeerA sent ReConfigure packet having OutgoingSSNResetRequestParameter  and IncomingSSNResetRequestParameter to PeerB and PeerB's responded with ReConfigure packet containing ReConfigurationResponseParameter(RCRP) and OutgoingSSNResetRequestParameter(OSRRP) having RCRP.ResponseSeqNumber equal to OSRRP.RequestSeqNumber. The missmatch occured when PeerB received response(ReConfigurationResponseParameter) from PeerA for the OutgoingSSNResetRequestParameter and tried to match with previously sent requests (parameters) from cache.